Update Exercício 003.py
Alterado o int para float, a fim de englobar mais números.

+++ Exercícios/Exercícios resolvidos/Exercício 003.py
@@ -1,7 +1,7 @@
 # Crie um programa que leia dois números e mostre a soma entre eles.
 
-numero1 = int(input('Digite um valor: '))
-numero2 = int(input('Digite outro valor: '))
+numero1 = float(input('Digite um valor: '))
+numero2 = float(input('Digite outro valor: '))
 soma = numero1 + numero2
 
 print(f'\nA soma entre {numero1} e {numero2} é igual a {soma}!')
